requirements.txt
python-telegram-bot[ext]
selenium
webdriver-manager
asyncio
